+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Factor out JFRProcessor into its own library
+- Merge `--preset` and `--config` into a single `--config` option
 
 ## [0.1.2] - 2026-01-05